Fixed indentation on plugins
