+++ _plugins/external-posts.rb
@@ -32,5 +32,4 @@ module ExternalPosts
       end
     end
   end
-
 end
