+++ _plugins/hide-custom-bibtex.rb
@@ -1,0 +1,15 @@
+module Jekyll
+  module HideCustomBibtex
+    def hideCustomBibtex(input)
+	    keywords = @context.registers[:site].config['filtered_bibtex_keywords']
+
+	    keywords.each do |keyword|
+		    input = input.gsub(/^.*\b#{keyword}\b *= *\{.*$\n/, '')
+	    end
+
+      return input
+    end
+  end
+end
+
+Liquid::Template.register_filter(Jekyll::HideCustomBibtex)
