--- _plugins/hideCustomBibtex.rb
@@ -1,15 +1,0 @@
- module Jekyll
-  module HideCustomBibtex
-    def hideCustomBibtex(input)
-	  keywords = @context.registers[:site].config['filtered_bibtex_keywords']
-
-	  keywords.each do |keyword|
-		input = input.gsub(/^.*\b#{keyword}\b *= *\{.*$\n/, '')
-	  end
-
-      return input
-    end
-  end
-end
-
-Liquid::Template.register_filter(Jekyll::HideCustomBibtex)
